refactor: change imports in typography
